Revert "Merge branch feature-psa-storage commits to master branch (#6)"
This reverts commit 483feeb75afeb03c693b388c293e76d5658b708c.
